+++ src/main/migration_booth/lib/entities/StatisticsCommunicator.cpp
@@ -1,18 +1,9 @@
 #include "StatisticsCommunicator.h"
+#include "FileNames.h"
 
-static const std::string STATS_FILE = "/bin/ls";
-static const std::string LOCK_STATS_FILE = "/tmp/statistics_shm_lock";
-static const char LETTER = 'A';
-
-enum FIELDS {
-    ALLOWED_RESIDENTS,
-    DETAINED_RESIDENTS,
-    ALLOWED_FOREIGNERS,
-    DEPORTED_FOREIGNERS
-};
-
-StatisticsCommunicator::StatisticsCommunicator() : stats_shm(STATS_FILE, LETTER),
-                                                   stats_shm_lock(LOCK_STATS_FILE) {}
+StatisticsCommunicator::StatisticsCommunicator() :
+        stats_shm(StatisticsSharedMemory::STATS_FILE, StatisticsSharedMemory::LETTER),
+        stats_shm_lock(StatisticsSharedMemory::LOCK_STATS_FILE) {}
 
 void StatisticsCommunicator::increment_field(size_t field) {
     stats_shm_lock.lock();
